# Conflicts:
#	features/recharge/common_digital/src/main/java/com/tokopedia/common_digital/atc/DigitalAddToCartViewModel.kt
#	features/recharge/common_topup_bills/src/test/java/com/tokopedia/common_digital/atc/DigitalAddToCartViewModelTest.kt
#	features/recharge/digital_product_detail/src/main/java/com/tokopedia/digital_product_detail/presentation/fragment/DigitalPDPDataPlanFragment.kt
#	features/recharge/digital_product_detail/src/main/java/com/tokopedia/digital_product_detail/presentation/viewmodel/DigitalPDPPulsaViewModel.kt
#	features/recharge/digital_product_detail/src/test/java/com/tokopedia/digital_product_detail/presentation/viewmodel/DigitalPDPDataPlanViewModelTest.kt
#	features/recharge/digital_product_detail/src/test/java/com/tokopedia/digital_product_detail/presentation/viewmodel/DigitalPDPDataPlanViewModelTestFixture.kt
#	features/recharge/digital_product_detail/src/test/java/com/tokopedia/digital_product_detail/presentation/viewmodel/DigitalPDPPulsaViewModelTest.kt
#	features/recharge/digital_product_detail/src/test/java/com/tokopedia/digital_product_detail/presentation/viewmodel/DigitalPDPPulsaViewModelTestFixture.kt

+++ features/user/chatbot/src/main/java/com/tokopedia/chatbot/attachinvoice/view/TransactionInvoiceBottomSheet.kt
@@ -31,7 +31,7 @@ class TransactionInvoiceBottomSheet : BottomSheetUnify(), TransactionInvoiceList
     private lateinit var touchViewPager: TouchViewPager
     private lateinit var transactionNotFoundButton: UnifyButton
     private lateinit var context: FragmentActivity
-    private var messageId: Int = 0
+    private var messageId: Long = 0
     private lateinit var listener: TransactionInvoiceBottomSheetListener
 
     init {
@@ -42,7 +42,7 @@ class TransactionInvoiceBottomSheet : BottomSheetUnify(), TransactionInvoiceList
 
     companion object {
         @JvmStatic
-        fun newInstance(context: FragmentActivity, messageId: Int,
+        fun newInstance(context: FragmentActivity, messageId: Long,
                         listener: TransactionInvoiceBottomSheetListener): TransactionInvoiceBottomSheet {
             return TransactionInvoiceBottomSheet().apply {
                 this.context = context
